Princess has no morals.
